Created data seed for import definitions
